backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
lxml>=5.0,<6.0
Pillow>=10.0,<11.0
